+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - Added Element content and style options in constructor.
 
+### Changes
+
+- Improved `Element` logs.
 
 ## 0.0.1 (Dec 6, 2021)